requirements.txt
fastapi==0.115.4
loguru==0.7.2
openai==1.54.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
python-multipart==0.0.17
tiktoken==0.8.0
uvicorn==0.32.0
